Remove network request for local client getRoles() call (#25)
We currently call the server for every getRoles() call, even for the local client. The register role call can take a while which can lead to this get call timing out. The change being checked in causes local role lookups to just use the results returned from the registerRoles() call. This ensures that we only make a single network request on behalf of the local client and we properly wait for that registration request to complete before proceding.